removed player class
